[GR-57817] Prepare JNI and Unsafe usage for JDK 24.
