+++ truffle/mx.truffle/suite.py
@@ -290,12 +290,19 @@ suite = {
         "sdk:POLYGLOT",
         "com.oracle.truffle.api.instrumentation",
         "com.oracle.truffle.api.exception",
+        "com.oracle.truffle.api.impl.asm",
       ],
       "requires" : [
         "java.logging",
         "jdk.management",
         "jdk.unsupported", # sun.misc.Unsafe
       ],
+      "requiresConcealed" : {
+        "java.base" : [
+          "jdk.internal.module",
+          "jdk.internal.access",
+        ],
+      },
       "annotationProcessors" : ["TRUFFLE_DSL_PROCESSOR"],
       "checkstyle" : "com.oracle.truffle.api",
       "javaCompliance" : "17+",
@@ -324,7 +331,6 @@ suite = {
           "jdk.vm.ci.code.stack",
         ],
         "java.base" : [
-          "jdk.internal.module",
           "jdk.internal.access",
         ],
       },
@@ -1285,7 +1291,7 @@ suite = {
       "graalCompilerSourceEdition": "ignore",
     },
 
-    "com.oracle.truffle.runtime.attach" : {
+    "com.oracle.truffle.attach" : {
       "subDir" : "src",
       "native" : "shared_lib",
       "deliverable" : "truffleattach",
@@ -1646,7 +1652,6 @@ suite = {
       "javaCompliance" : "17+",
       "dependencies" : [
         "com.oracle.truffle.runtime",
-        "TRUFFLE_RUNTIME_ATTACH_RESOURCES",
       ],
       "distDependencies" : [
         "sdk:JNIUTILS",
@@ -1762,6 +1767,7 @@ suite = {
         "com.oracle.truffle.host",
         "com.oracle.truffle.api.staticobject",
         "TRUFFLE_API_VERSION",
+        "TRUFFLE_ATTACH_RESOURCES",
       ],
       "distDependencies" : [
         "sdk:COLLECTIONS",
@@ -1786,7 +1792,7 @@ suite = {
       #}
     },
 
-    "TRUFFLE_RUNTIME_ATTACH_RESOURCES" : {
+    "TRUFFLE_ATTACH_RESOURCES" : {
       "type" : "dir",
       "platformDependent" : True,
       "hashEntry" :  "META-INF/resources/engine/libtruffleattach/<os>/<arch>/sha256",
@@ -1800,9 +1806,9 @@ suite = {
           "windows-aarch64",
       ],
       "layout" : {
-        "META-INF/resources/engine/libtruffleattach/<os>/<arch>/bin/" : "dependency:com.oracle.truffle.runtime.attach",
+        "META-INF/resources/engine/libtruffleattach/<os>/<arch>/bin/" : "dependency:com.oracle.truffle.attach",
       },
-      "description" : "Contains a library to provide access for the Truffle runtime to JVMCI.",
+      "description" : "Contains a library to provide access for the Truffle to JDK internal classes.",
       "maven": False,
       "graalCompilerSourceEdition": "ignore",
     },
