+++ truffle/src/com.oracle.truffle.nfi.backend.libffi/src/com/oracle/truffle/nfi/backend/libffi/LibFFILanguage.java
@@ -58,7 +58,7 @@ import com.oracle.truffle.nfi.backend.spi.NFIState;
 import com.oracle.truffle.nfi.backend.spi.types.NativeSimpleType;
 
 @TruffleLanguage.Registration(id = "internal/nfi-native", name = "nfi-native", version = "0.1", characterMimeTypes = LibFFILanguage.MIME_TYPE, internal = true, //
-                services = NFIBackendFactory.class, contextPolicy = ContextPolicy.SHARED, internalResources = LibNFIResource.class)
+                services = NFIBackendFactory.class, contextPolicy = ContextPolicy.SHARED, internalResources = LibNFIResource.class, enableNativeAccess = true)
 public class LibFFILanguage extends TruffleLanguage<LibFFIContext> {
 
     public static final String MIME_TYPE = "trufflenfi/native";
