+++ truffle/src/com.oracle.truffle.polyglot/src/com/oracle/truffle/polyglot/ModuleUtils.java
@@ -40,19 +40,71 @@
  */
 package com.oracle.truffle.polyglot;
 
+import java.io.IOException;
+import java.lang.invoke.MethodHandle;
+import java.lang.invoke.MethodHandles;
+import java.lang.invoke.MethodType;
+import java.lang.module.Configuration;
 import java.lang.module.ModuleDescriptor;
+import java.lang.module.ModuleFinder;
+import java.lang.module.ModuleReader;
+import java.lang.module.ModuleReference;
+import java.lang.reflect.Method;
+import java.net.URI;
+import java.nio.ByteBuffer;
+import java.nio.file.Path;
 import java.util.ArrayDeque;
+import java.util.Collections;
 import java.util.Deque;
+import java.util.EnumSet;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
+import java.util.function.Consumer;
+import java.util.function.Predicate;
+import java.util.stream.Stream;
 
+import com.oracle.truffle.api.InternalResource;
 import com.oracle.truffle.api.Truffle;
+import com.oracle.truffle.api.impl.Accessor;
+import com.oracle.truffle.api.impl.asm.ClassWriter;
+import com.oracle.truffle.api.impl.asm.MethodVisitor;
+import com.oracle.truffle.api.impl.asm.Opcodes;
+import jdk.internal.access.SharedSecrets;
+import jdk.internal.module.Modules;
 
 final class ModuleUtils {
+
+    /*
+     * We use reflective access to addEnableAccess and addEnableNativeAccessToAllUnnamed
+     * to make Truffle compilable on JDK-17.
+     */
+    private static final Method ADD_ENABLE_NATIVE_ACCESS;
+    private static final Method ADD_ENABLE_NATIVE_ACCESS_TO_ALL_UNNAMED;
+    static {
+        if (Runtime.version().feature() >= 21) {
+            try {
+                Class<?> langAccessClass = Class.forName("jdk.internal.access.JavaLangAccess");
+                ADD_ENABLE_NATIVE_ACCESS = langAccessClass.getDeclaredMethod("addEnableNativeAccess", Module.class);
+                ADD_ENABLE_NATIVE_ACCESS_TO_ALL_UNNAMED = langAccessClass.getDeclaredMethod("addEnableNativeAccessToAllUnnamed");
+            } catch (ReflectiveOperationException re) {
+                throw new InternalError(re);
+            }
+        } else {
+            // On JDK-17 do nothing
+            ADD_ENABLE_NATIVE_ACCESS = null;
+            ADD_ENABLE_NATIVE_ACCESS_TO_ALL_UNNAMED = null;
+        }
+    }
+
+    private static volatile Accessor.ModulesAccessor modulesAccessor;
+
+
+    private ModuleUtils() {
+    }
 
     static void exportTo(Module clientModule) {
         if (!isExportedTo(clientModule)) {
@@ -68,9 +120,61 @@ final class ModuleUtils {
         if (isExportedTo(clientModule)) {
             return;
         }
-
-        ModuleLayer layer = clientModule.getLayer();
         Module truffleModule = Truffle.class.getModule();
+        forEach(clientModule, EnumSet.of(Edge.READS, Edge.USES), (m) -> m != truffleModule && m.canRead(truffleModule), ModuleUtils::exportFromTo);
+    }
+
+    static void enableNativeAccess(Module clientModule) {
+        forEach(clientModule, EnumSet.of(Edge.READS), (m) -> true, (m) -> {
+            if (m.isNamed()) {
+                getModulesAccessor().addEnableNativeAccess(m);
+            } else {
+                getModulesAccessor().addEnableNativeAccessToAllUnnamed();
+            }
+        });
+    }
+
+    @SuppressWarnings("restricted")
+    static Accessor.ModulesAccessor getModulesAccessor() {
+        Accessor.ModulesAccessor result = modulesAccessor;
+        if (result == null) {
+            synchronized (ModuleUtils.class) {
+                result = modulesAccessor;
+                if (result == null) {
+                    String attachLibPath = System.getProperty("truffle.attach.library");
+                    if (attachLibPath == null) {
+                        try {
+                            Path truffleAttachRoot = InternalResourceCache.installRuntimeResource(new LibTruffleAttachResource());
+                            Path libAttach = truffleAttachRoot.resolve("bin").resolve(System.mapLibraryName("truffleattach"));
+                            attachLibPath = libAttach.toString();
+                        } catch (IOException ioe) {
+                            throw new InternalError(ioe);
+                        }
+                    }
+                    System.load(attachLibPath);
+                    if (Accessor.ModulesAccessor.class.getModule().isNamed()) {
+                        result = new DirectImpl(Accessor.ModulesAccessor.class);
+                    } else {
+                        try {
+                            result = new IsolatedImpl(Accessor.ModulesAccessor.class);
+                        } catch (ReflectiveOperationException e) {
+                            throw new InternalError(e);
+                        }
+                    }
+                    Module javaBase = ModuleLayer.boot().findModule("java.base").orElseThrow();
+                    addExports0(javaBase, "jdk.internal.module", result.getTargetModule());
+                    result.addExports(javaBase, "jdk.internal.access", result.getTargetModule());
+                    modulesAccessor = result;
+                }
+            }
+        }
+        return result;
+    }
+
+    private static native void addExports0(Module m1, String pn, Module m2);
+
+    private static void forEach(Module rootModule, Set<Edge> edges, Predicate<? super Module> filter, Consumer<? super Module> action) {
+        ModuleLayer layer = rootModule.getLayer();
         Set<Module> targetModules = new HashSet<>();
         Deque<Module> todo = new ArrayDeque<>();
         /*
@@ -78,11 +182,11 @@ final class ModuleUtils {
          * visited modules to detect cycles.
          */
         Set<Module> visited = new HashSet<>();
-        todo.add(clientModule);
+        todo.add(rootModule);
         Map<String, Set<Module>> serviceDictionary = null;
         while (!todo.isEmpty()) {
             Module module = todo.removeFirst();
-            if (visited.add(module) && Objects.equals(module.getLayer(), layer) && readsTruffleModule(truffleModule, module)) {
+            if (visited.add(module) && Objects.equals(module.getLayer(), layer) && filter.test(module)) {
                 targetModules.add(module);
                 ModuleDescriptor descriptor = module.getDescriptor();
                 if (descriptor == null) {
@@ -105,37 +209,37 @@ final class ModuleUtils {
                      * Named module with a module descriptor: Export transitively to all modules
                      * required by the named module.
                      */
-                    for (ModuleDescriptor.Requires requires : descriptor.requires()) {
-                        Module requiredModule = findModule(layer, requires);
-                        if (requiredModule != null) {
-                            todo.add(requiredModule);
+                    if (edges.contains(Edge.READS)) {
+                        for (ModuleDescriptor.Requires requires : descriptor.requires()) {
+                            Module requiredModule = findModule(layer, requires);
+                            if (requiredModule != null) {
+                                todo.add(requiredModule);
+                            }
                         }
                     }
-                    // Open also to modules providing a service consumed by the module.
-                    Set<String> usedServices = descriptor.uses();
-                    if (!usedServices.isEmpty()) {
-                        if (serviceDictionary == null) {
-                            serviceDictionary = new HashMap<>();
-                            for (Module m : layer.modules()) {
-                                if (readsTruffleModule(truffleModule, m)) {
-                                    for (ModuleDescriptor.Provides provides : m.getDescriptor().provides()) {
-                                        serviceDictionary.computeIfAbsent(provides.service(), (k) -> new HashSet<>()).add(m);
+                    if (edges.contains(Edge.USES)) {
+                        // Open also to modules providing a service consumed by the module.
+                        Set<String> usedServices = descriptor.uses();
+                        if (!usedServices.isEmpty()) {
+                            if (serviceDictionary == null) {
+                                serviceDictionary = new HashMap<>();
+                                for (Module m : layer.modules()) {
+                                    if (filter.test(m)) {
+                                        for (ModuleDescriptor.Provides provides : m.getDescriptor().provides()) {
+                                            serviceDictionary.computeIfAbsent(provides.service(), (k) -> new HashSet<>()).add(m);
+                                        }
                                     }
                                 }
                             }
-                        }
-                        for (String service : usedServices) {
-                            todo.addAll(serviceDictionary.getOrDefault(service, Set.of()));
+                            for (String service : usedServices) {
+                                todo.addAll(serviceDictionary.getOrDefault(service, Set.of()));
+                            }
                         }
                     }
                 }
             }
         }
-        targetModules.forEach(ModuleUtils::exportFromTo);
-    }
-
-    private static boolean readsTruffleModule(Module truffleModule, Module otherModule) {
-        return otherModule != truffleModule && otherModule.canRead(truffleModule);
+        targetModules.forEach(action);
     }
 
     private static Module findModule(ModuleLayer layer, ModuleDescriptor.Requires requires) {
@@ -173,4 +277,338 @@ final class ModuleUtils {
         }
     }
 
+    private enum Edge {
+        READS,
+        USES
+    }
+
+    /*
+     * When using a named module we do not need to do anything and we can just call the Modules
+     * class. We used a qualified export to export it to this class.
+     */
+    private static final class DirectImpl extends Accessor.ModulesAccessor {
+
+        private final Class<?> baseClass;
+
+        DirectImpl(Class<?> baseClass) {
+            if (!baseClass.getModule().isNamed()) {
+                throw new IllegalStateException("");
+            }
+            this.baseClass = baseClass;
+        }
+
+        @Override
+        public void addExports(Module base, String p, Module target) {
+            Modules.addExports(base, p, target);
+        }
+
+        @Override
+        public void addExportsToAllUnnamed(Module base, String p) {
+            Modules.addExportsToAllUnnamed(base, p);
+        }
+
+        @Override
+        public void addOpens(Module base, String p, Module target) {
+            Modules.addOpens(base, p, target);
+        }
+
+        @Override
+        public void addOpensToAllUnnamed(Module base, String p) {
+            Modules.addOpensToAllUnnamed(base, p);
+        }
+
+        @Override
+        public void addEnableNativeAccess(Module module) {
+            if (ADD_ENABLE_NATIVE_ACCESS != null) {
+                try {
+                    ADD_ENABLE_NATIVE_ACCESS.invoke(SharedSecrets.getJavaLangAccess(), module);
+                } catch (ReflectiveOperationException re) {
+                    throw new InternalError(re);
+                }
+            }
+        }
+
+        @Override
+        public void addEnableNativeAccessToAllUnnamed() {
+            if (ADD_ENABLE_NATIVE_ACCESS_TO_ALL_UNNAMED != null) {
+                try {
+                    ADD_ENABLE_NATIVE_ACCESS_TO_ALL_UNNAMED.invoke(SharedSecrets.getJavaLangAccess());
+                } catch (ReflectiveOperationException re) {
+                    throw new InternalError(re);
+                }
+            }
+        }
+
+        @Override
+        public Module getTargetModule() {
+            return baseClass.getModule();
+        }
+    }
+
+    /*
+     * When using a named module we do not need to do anything and we can just call the Modules
+     * class. We used a qualified export to export it to this class.
+     */
+    private static final class IsolatedImpl extends Accessor.ModulesAccessor {
+
+        private final MethodHandle addExports;
+        private final MethodHandle addExportsToAllUnnamed;
+        private final MethodHandle addOpens;
+        private final MethodHandle addOpensToAllUnnamed;
+        private final MethodHandle addEnableNativeAccess;
+        private final MethodHandle addEnableNativeAccessToAllUnnamed;
+        private final Module targetModule;
+
+        IsolatedImpl(Class<?> baseClass) throws ReflectiveOperationException {
+            final String moduleName = "org.graalvm.truffle.generated";
+            final String targetPackage = baseClass.getPackageName() + ".generated";
+            final String className = targetPackage + ".GeneratedModules";
+            final String binaryClassName = className.replace('.', '/');
+
+            ClassWriter cw = new ClassWriter(0);
+            cw.visit(Opcodes.V1_8, Opcodes.ACC_PUBLIC + Opcodes.ACC_SUPER, binaryClassName, null, "java/lang/Object", null);
+
+            MethodVisitor constructor = cw.visitMethod(Opcodes.ACC_PUBLIC, "<init>", "()V", null, null);
+            constructor.visitCode();
+            constructor.visitVarInsn(Opcodes.ALOAD, 0);
+            constructor.visitMethodInsn(Opcodes.INVOKESPECIAL, "java/lang/Object", "<init>", "()V", false);
+            constructor.visitInsn(Opcodes.RETURN);
+            constructor.visitMaxs(1, 1);
+            constructor.visitEnd();
+
+            MethodVisitor mv1 = cw.visitMethod(Opcodes.ACC_PUBLIC + Opcodes.ACC_STATIC, "addExports",
+                    "(Ljava/lang/Module;Ljava/lang/String;Ljava/lang/Module;)V", null, null);
+            mv1.visitCode();
+            mv1.visitVarInsn(Opcodes.ALOAD, 0); // Load first argument (Module base)
+            mv1.visitVarInsn(Opcodes.ALOAD, 1); // Load second argument (String p)
+            mv1.visitVarInsn(Opcodes.ALOAD, 2); // Load third argument (Module target)
+            mv1.visitMethodInsn(Opcodes.INVOKESTATIC, "jdk/internal/module/Modules", "addExports",
+                    "(Ljava/lang/Module;Ljava/lang/String;Ljava/lang/Module;)V", false);
+            mv1.visitInsn(Opcodes.RETURN);
+            mv1.visitMaxs(3, 3);
+            mv1.visitEnd();
+
+            MethodVisitor mv2 = cw.visitMethod(Opcodes.ACC_PUBLIC + Opcodes.ACC_STATIC, "addExportsToAllUnnamed",
+                    "(Ljava/lang/Module;Ljava/lang/String;)V", null, null);
+            mv2.visitCode();
+            mv2.visitVarInsn(Opcodes.ALOAD, 0); // Load first argument (Module target)
+            mv2.visitVarInsn(Opcodes.ALOAD, 1); // Load second argument (String p)
+            mv2.visitMethodInsn(Opcodes.INVOKESTATIC, "jdk/internal/module/Modules", "addExportsToAllUnnamed",
+                    "(Ljava/lang/Module;Ljava/lang/String;)V", false);
+            mv2.visitInsn(Opcodes.RETURN);
+            mv2.visitMaxs(2, 2);
+            mv2.visitEnd();
+
+            MethodVisitor mv3 = cw.visitMethod(Opcodes.ACC_PUBLIC + Opcodes.ACC_STATIC, "addOpens",
+                    "(Ljava/lang/Module;Ljava/lang/String;Ljava/lang/Module;)V", null, null);
+            mv3.visitCode();
+            mv3.visitVarInsn(Opcodes.ALOAD, 0); // Load first argument (Module base)
+            mv3.visitVarInsn(Opcodes.ALOAD, 1); // Load second argument (String p)
+            mv3.visitVarInsn(Opcodes.ALOAD, 2); // Load third argument (Module target)
+            mv3.visitMethodInsn(Opcodes.INVOKESTATIC, "jdk/internal/module/Modules", "addOpens",
+                    "(Ljava/lang/Module;Ljava/lang/String;Ljava/lang/Module;)V", false);
+            mv3.visitInsn(Opcodes.RETURN);
+            mv3.visitMaxs(3, 3);
+            mv3.visitEnd();
+
+            MethodVisitor mv4 = cw.visitMethod(Opcodes.ACC_PUBLIC + Opcodes.ACC_STATIC, "addOpensToAllUnnamed",
+                    "(Ljava/lang/Module;Ljava/lang/String;)V", null, null);
+            mv4.visitCode();
+            mv4.visitVarInsn(Opcodes.ALOAD, 0); // Load first argument (Module target)
+            mv4.visitVarInsn(Opcodes.ALOAD, 1); // Load second argument (String p)
+            mv4.visitMethodInsn(Opcodes.INVOKESTATIC, "jdk/internal/module/Modules", "addOpensToAllUnnamed",
+                    "(Ljava/lang/Module;Ljava/lang/String;)V", false);
+            mv4.visitInsn(Opcodes.RETURN);
+            mv4.visitMaxs(2, 2);
+            mv4.visitEnd();
+
+            MethodVisitor mv5 = cw.visitMethod(Opcodes.ACC_PUBLIC + Opcodes.ACC_STATIC, "addEnableNativeAccess",
+                    "(Ljava/lang/Module;)V", null, null);
+            mv5.visitCode();
+            if (ADD_ENABLE_NATIVE_ACCESS != null) {
+                mv5.visitMethodInsn(Opcodes.INVOKESTATIC, "jdk/internal/access/SharedSecrets", "getJavaLangAccess",
+                        "()Ljdk/internal/access/JavaLangAccess;", false);
+                mv5.visitVarInsn(Opcodes.ALOAD, 0); // Load first argument (Module module)
+                mv5.visitMethodInsn(Opcodes.INVOKEINTERFACE, "jdk/internal/access/JavaLangAccess", "addEnableNativeAccess",
+                        "(Ljava/lang/Module;)Ljava/lang/Module;", true);
+            }
+            mv5.visitInsn(Opcodes.RETURN);
+            mv5.visitMaxs(2, 2);
+            mv5.visitEnd();
+
+            MethodVisitor mv6 = cw.visitMethod(Opcodes.ACC_PUBLIC + Opcodes.ACC_STATIC, "addEnableNativeAccessToAllUnnamed",
+                    "()V", null, null);
+            mv6.visitCode();
+            if (ADD_ENABLE_NATIVE_ACCESS_TO_ALL_UNNAMED != null) {
+                mv6.visitMethodInsn(Opcodes.INVOKESTATIC, "jdk/internal/access/SharedSecrets", "getJavaLangAccess",
+                        "()Ljdk/internal/access/JavaLangAccess;", false);
+                mv6.visitMethodInsn(Opcodes.INVOKEINTERFACE, "jdk/internal/access/JavaLangAccess", "addEnableNativeAccessToAllUnnamed",
+                        "()V", true);
+            }
+            mv6.visitInsn(Opcodes.RETURN);
+            mv6.visitMaxs(1, 1);
+            mv6.visitEnd();
+
+            cw.visitEnd();
+
+            byte[] classBytes = cw.toByteArray();
+
+            // Create a ModuleDescriptor for the new module
+            ModuleDescriptor descriptor = ModuleDescriptor.newModule(moduleName).exports(targetPackage).build();
+
+            // Create a ModuleFinder that finds the module and class
+            ModuleFinder finder = new ModuleFinder() {
+
+                public Optional<ModuleReference> find(String name) {
+                    if (name.equals(moduleName)) {
+                        return Optional.of(new ModuleReference(descriptor, null) {
+                            @Override
+                            @SuppressWarnings("hiding")
+                            public ModuleReader open() throws IOException {
+                                return new ModuleReader() {
+                                    @Override
+                                    public Optional<ByteBuffer> read(String name) throws IOException {
+                                        if (name.equals(binaryClassName + ".class")) {
+                                            return Optional.of(ByteBuffer.wrap(classBytes));
+                                        }
+                                        return Optional.empty();
+                                    }
+
+                                    @Override
+                                    public void close() throws IOException {
+                                    }
+
+                                    public Optional<URI> find(String name) throws IOException {
+                                        return Optional.empty();
+                                    }
+
+                                    public Stream<String> list() throws IOException {
+                                        return Stream.empty();
+                                    }
+                                };
+                            }
+                        });
+                    }
+                    return Optional.empty();
+                }
+
+                @Override
+                public Set<ModuleReference> findAll() {
+                    return Collections.singleton(find(moduleName).get());
+                }
+            };
+
+            // Resolve the configuration for the new module layer based on the boot layer
+            ModuleLayer bootLayer = ModuleLayer.boot();
+            Configuration cf = bootLayer.configuration().resolve(finder, ModuleFinder.of(), Collections.singleton(moduleName));
+
+            // Define a new module layer with the correct parent layer and class loader
+            ModuleLayer layer = bootLayer.defineModulesWithOneLoader(cf, ClassLoader.getSystemClassLoader());
+
+            // Load the class from the module layer
+            Class<?> generatedClass = layer.findLoader(moduleName).loadClass(className);
+            this.targetModule = generatedClass.getModule();
+
+            MethodHandles.Lookup l = MethodHandles.lookup();
+            l.accessClass(generatedClass);
+            this.addExports = l.findStatic(generatedClass, "addExports", MethodType.methodType(void.class, Module.class, String.class, Module.class));
+            this.addExportsToAllUnnamed = l.findStatic(generatedClass, "addExportsToAllUnnamed", MethodType.methodType(void.class, Module.class, String.class));
+            this.addOpens = l.findStatic(generatedClass, "addOpens", MethodType.methodType(void.class, Module.class, String.class, Module.class));
+            this.addOpensToAllUnnamed = l.findStatic(generatedClass, "addOpensToAllUnnamed", MethodType.methodType(void.class, Module.class, String.class));
+            this.addEnableNativeAccess = l.findStatic(generatedClass, "addEnableNativeAccess", MethodType.methodType(void.class, Module.class));
+            this.addEnableNativeAccessToAllUnnamed = l.findStatic(generatedClass, "addEnableNativeAccessToAllUnnamed", MethodType.methodType(void.class));
+        }
+
+        @Override
+        public Module getTargetModule() {
+            return targetModule;
+        }
+
+        @Override
+        public void addExports(Module base, String p, Module target) {
+            try {
+                addExports.invokeExact(base, p, target);
+            } catch (Throwable e) {
+                throw new InternalError(e);
+            }
+        }
+
+        @Override
+        public void addExportsToAllUnnamed(Module base, String p) {
+            try {
+                addExportsToAllUnnamed.invokeExact(base, p);
+            } catch (Throwable e) {
+                throw new InternalError(e);
+            }
+        }
+
+        @Override
+        public void addOpens(Module base, String p, Module target) {
+            try {
+                addOpens.invokeExact(base, p, target);
+            } catch (Throwable e) {
+                throw new InternalError(e);
+            }
+        }
+
+        @Override
+        public void addOpensToAllUnnamed(Module base, String p) {
+            try {
+                addOpensToAllUnnamed.invokeExact(base, p);
+            } catch (Throwable e) {
+                throw new InternalError(e);
+            }
+        }
+
+        @Override
+        public void addEnableNativeAccess(Module module) {
+            try {
+                addEnableNativeAccess.invokeExact(module);
+            } catch (Throwable e) {
+                throw new InternalError(e);
+            }
+        }
+
+        @Override
+        public void addEnableNativeAccessToAllUnnamed() {
+            try {
+                addEnableNativeAccessToAllUnnamed.invokeExact();
+            } catch (Throwable e) {
+                throw new InternalError(e);
+            }
+        }
+    }
+
+    @InternalResource.Id(value = LibTruffleAttachResource.ID, componentId = "engine", optional = true)
+    final static class LibTruffleAttachResource implements InternalResource {
+
+        static final String ID = "libtruffleattach";
+
+        LibTruffleAttachResource() {
+        }
+
+        @Override
+        public void unpackFiles(Env env, Path targetDirectory) throws IOException {
+            if (env.inNativeImageBuild() && !env.inContextPreinitialization()) {
+                // The truffleattach library is not needed in native-image.
+                return;
+            }
+            Path base = basePath(env);
+            env.unpackResourceFiles(base.resolve("files"), targetDirectory, base);
+        }
+
+        @Override
+        public String versionHash(Env env) {
+            try {
+                Path base = basePath(env);
+                return env.readResourceLines(base.resolve("sha256")).get(0);
+            } catch (IOException ioe) {
+                throw new InternalError(ioe);
+            }
+        }
+
+        private static Path basePath(Env env) {
+            return Path.of("META-INF", "resources", "engine", ID, env.getOS().toString(), env.getCPUArchitecture().toString());
+        }
+    }
 }
